requirements.txt
coverage==4.5.1
Django==2.1
django-ordered-model==2.1.0
factory-boy==2.11.1
Pillow==5.2.0
